Split drawing tests
